MEGA BUNDLE COMMIT - long awaited transaction detail view redesign - transaction comment function repaired, that was broken since breadwallet fork - send confirmation view redesigned - green glowing shield after security steps are fulfilled - touchID prompt at application startup fixed - balanceView collapsable - balanceView toggable again - qrcode encoding fix - little sync stall issue - fadeView is now blue blurView - refactoring & several bugfixes

+++ breadwallet/src/BRCore.swift
@@ -567,6 +567,9 @@ class BRPeerManager {
         },
         { (info) in // txStatusUpdate
             guard let info = info else { return }
+            // ToDo: After WipeWallet sometimes we get EXC_BAD_ACCESS (swift_unknownRetain)
+            // That probably happens if memory was released before we access the takeUnretainedValue()
+            // Is there a way to fix that?
             Unmanaged<BRPeerManager>.fromOpaque(info).takeUnretainedValue().listener.txStatusUpdate()
         },
         { (info, replace, blocks, blocksCount) in // saveBlocks
@@ -690,7 +693,7 @@ class BRPeerManager {
         return BRPeerManagerRelayCount(cPtr, forTxHash)
     }
 
-    func setFixedPeer(address: Int, port: Int) {
+    func setFixedPeer(address: Int, port: UInt16) {
         if address != 0 {
             var newAddress = UInt128()
             newAddress.u16.5 = 0xffff
